fixes issue where backing up db reported a nullpointer exception. split out the delegate to run all of the proper statements for a backup database helper
